Add rds, elb, web security groups
A security group acts as a virtual firewall that controls the traffic
for one or more instances. When you launch an instance, you associate
one or more security groups with the instance. You add rules to each
security group that allow traffic to or from its associated
instances. You can modify the rules for a security group at any time;
the new rules are automatically applied to all instances that are
associated with the security group. When we decide whether to allow
traffic to reach an instance, we evaluate all the rules from all the
security groups that are associated with the instance.

WEB-SG: Allows SSH from whitelist_ip and http(s) from ELB-SG
RDS-SG: Allows MYSQL Conections from any instance with the WEB-SG
elb-sg: Allows HTTP(s) from everybody!